[Git] Simplify .mailmap
This commit replaces the external reference to `Git Common.sublime-syntax#email` by an extended version of the existing lazy email pattern, defined in the syntax itself as most of the email parts are already there.

Furthermore it removes matching `[ \t]*` in front of the "proper name" as it does not have any effect.